fix: Fix TypedArray element conversion semantics to follow QuickJS ToInt32/ToUint8Clamp behavior.

+++ src/main/java/com/caoccao/qjs4j/core/JSFloat16Array.java
@@ -103,7 +103,7 @@ public final class JSFloat16Array extends JSTypedArray {
     public void setElement(int index, double value) {
         checkIndex(index);
         ByteBuffer buf = getByteBuffer();
-        short halfFloat = Float16.toHalf((float) value);
+        short halfFloat = Float16.toHalf(value);
         buf.putShort(index * BYTES_PER_ELEMENT, halfFloat);
     }
